feat: CRUD, paginação e ordenação

+++ src/schemas/movie.schema.ts
@@ -4,8 +4,8 @@ export const moviesSchema = z.object({
     id: z.number().positive(),
     name: z.string().max(50).nonempty(),
     description: z.string().nullish(),
-    duration: z.number().positive(),
-    price: z.number().int().positive(),
+    duration: z.number().positive().int(),
+    price: z.number().positive().int()
 });
 
 export const moviesCreateSchema = moviesSchema.omit({ id: true})
